requirements.txt
revolt @ git+https://github.com/revoltchat/revolt.py
python-dotenv